server/requirements.txt
alembic==1.12.1 ; python_version >= "3.12" and python_version < "4.0"
annotated-types==0.6.0 ; python_version >= "3.12" and python_version < "4.0"
anyio==3.7.1 ; python_version >= "3.12" and python_version < "4.0"
asyncpg==0.29.0 ; python_version >= "3.12" and python_version < "4.0"
bcrypt==4.0.1 ; python_version >= "3.12" and python_version < "4.0"
cffi==1.16.0 ; python_version >= "3.12" and python_version < "4.0"
cryptography==41.0.5 ; python_version >= "3.12" and python_version < "4.0"
dnspython==2.4.2 ; python_version >= "3.12" and python_version < "4.0"
email-validator==2.1.0.post1 ; python_version >= "3.12" and python_version < "4.0"
fastapi==0.104.1 ; python_version >= "3.12" and python_version < "4.0"
greenlet==3.0.1 ; python_version >= "3.12" and python_version < "4.0" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32")
idna==3.4 ; python_version >= "3.12" and python_version < "4.0"
mako==1.2.4 ; python_version >= "3.12" and python_version < "4.0"
markupsafe==2.1.3 ; python_version >= "3.12" and python_version < "4.0"
passlib[bcrypt]==1.7.4 ; python_version >= "3.12" and python_version < "4.0"
pycparser==2.21 ; python_version >= "3.12" and python_version < "4.0"
pydantic-core==2.10.1 ; python_version >= "3.12" and python_version < "4.0"
pydantic-settings==2.0.3 ; python_version >= "3.12" and python_version < "4.0"
pydantic==2.4.2 ; python_version >= "3.12" and python_version < "4.0"
pydantic[dotenv,email]==2.4.2 ; python_version >= "3.12" and python_version < "4.0"
pyjwt[crypto]==2.8.0 ; python_version >= "3.12" and python_version < "4.0"
python-dotenv==1.0.0 ; python_version >= "3.12" and python_version < "4.0"
python-multipart==0.0.6 ; python_version >= "3.12" and python_version < "4.0"
sniffio==1.3.0 ; python_version >= "3.12" and python_version < "4.0"
sqlalchemy==2.0.23 ; python_version >= "3.12" and python_version < "4.0"
starlette==0.27.0 ; python_version >= "3.12" and python_version < "4.0"
typing-extensions==4.8.0 ; python_version >= "3.12" and python_version < "4.0"
uvicorn==0.24.0